requirements.txt
asgiref==3.5.0
coverage==6.3.2
Django==4.0.3
django-crispy-forms==1.14.0
django-multiselectfield==0.1.12
django-owner-required==2020.7.1
django-phonenumber-field==6.1.0
gunicorn==20.1.0
phonenumbers==8.12.45
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
